fix(ui): accumulate streaming content instead of replacing it (#860)
when a non-final status update comes with text content we currently
replace the streaming content with the current text content. This
results in scenarios where only the agent's latest text content would be
streamed. Ideally before the final status update we want to stream all
of the text content instead of the latest text content so the agents
thinking process is streamed in its entirety. This PR fixes this issue
by accumulating the streaming content prior to a final status update.

resolves https://github.com/kagent-dev/kagent/issues/852

Below are two short recordings demonstrating the issue and the results
after the fix

[before-fix.zip](https://github.com/user-attachments/files/22145256/before-fix.zip)

[after-fix.mov.zip](https://github.com/user-attachments/files/22145278/after-fix.mov.zip)

Signed-off-by: JM Huibonhoa <jm.huibonhoa@solo.io>