feat: migrate kmcp cli into kagent cli (#953)
This PR merges the kmcp cli commands ino the kagent cli with the
available commands shown below.
```
Usage:
  kagent mcp [command]

Available Commands:
  add-tool    Add a new MCP tool to your project
  build       Build MCP server as a Docker image
  deploy      Deploy MCP server to Kubernetes
  init        Initialize a new MCP server project
  run         Run MCP server locally
  secrets     Manage project secrets
  ``` 
A user using the kagent cli is now able to scaffold an mcp server in `golang`, `python`, `typescript` and `java` and using the CLI will have the ability to run the mcp server locally or deploy it to their kubernetes environment. 

**Testing**
- Tested all commands above in my kind cluster for each supported framework ie: `golang`, `python`, `typescript` and `java`

---------

Signed-off-by: JM Huibonhoa <jm.huibonhoa@solo.io>